@@ -3,8 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@chakra-ui/icons": "^1.0.16",
-    "@chakra-ui/react": "^1.6.10",
+    "@chakra-ui/icons": "^2",
+    "@chakra-ui/react": "^2",
     "@emotion/react": "^11",
     "@emotion/styled": "^11",
     "axios": "^0.22.0",
@@ -26,8 +26,8 @@
     "morgan": "^1.10.0",
     "nanoid": "^4.0.0",
     "nodemon": "^2.0.13",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18",
+    "react-dom": "^18",
     "react-horizontal-scrolling-menu": "^2.5.1",
     "react-icons": "^4.3.1",
     "react-query": "^3.29.0",
