@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
@@ -11,7 +11,9 @@ import {Suspense} from 'react';
 
 const queryClient = new QueryClient()
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(
     <React.StrictMode>
         <RecoilRoot>
             <QueryClientProvider client={queryClient}>
@@ -23,8 +25,7 @@ ReactDOM.render(
             </QueryClientProvider>
         </RecoilRoot>
     </React.StrictMode>,
-    document.getElementById('root')
-);
+)
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
